pyproject.toml
[tool.poetry]
name = "stream-ml-housing-lab"
version = "0.1.0"
description = "Production-grade housing price ML pipeline following MLOps best practices"
authors = ["Hanish Moola"]
license = "Apache-2.0"
readme = "README.md"
packages = [{ include = "src" }]

[tool.poetry.dependencies]
python = "^3.10"
mlflow = "^2.9.0"
scikit-learn = "^1.3.0"
pandas = "^2.1.0"
numpy = "^1.25.0"
joblib = "^1.3.0"
pyyaml = "^6.0.1"
xgboost = "^1.7.6"
tensorflow = "^2.13.0"
fastapi = "^0.104.0"
uvicorn = "^0.23.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
black = "^23.9.0"
ruff = "^0.1.3"
mypy = "^1.5.1"

[tool.poetry.scripts]
feature-engineering = "src.feature_engineering:main"
train-model = "src.train:main"
evaluate-model = "src.evaluate:main"
predict-model = "src.predict:main"
offline-train = "src.workflows.offline_train:main"
serve-api = "src.api.app:main"
stream-infer = "src.workflows.stream_infer:main"

[build-system]
requires = ["poetry-core>=1.5.0"]
build-backend = "poetry.core.masonry.api"
